Move segment-related struct into its own header.
